big refactor to break up cli file into modular components
