+++ ue/cli.py
@@ -1,30 +1,7 @@
 """Command-line interface."""
 
 import click
-from rich.console import Console
-
-console = Console()
-
-LOGO = "[green][■■■□□] utility-explorer[/green]"
-
-
-def print_logo():
-    """Print the utility-explorer logo."""
-    console.print(LOGO)
-
-
-def get_effective_date():
-    """Get the 'effective' date, treating 2am as the day boundary.
-
-    If it's between midnight and 2am, returns yesterday's date.
-    This matches natural sleep cycles - 12:41am still feels like 'tonight'.
-    """
-    from datetime import datetime, timedelta
-
-    now = datetime.now()
-    if now.hour < 2:
-        return (now - timedelta(days=1)).date()
-    return now.date()
+from ue.utils.display import console, print_logo
 
 
 @click.group()
@@ -98,1345 +75,47 @@ def setup():
     console.print("\nRun [cyan]ue sync[/cyan] to fetch data from all sources.")
 
 
-def run_sync(days: int = 7, quiet: bool = False):
-    """Run sync and record timestamp. Used by sync command and auto-sync."""
-    from datetime import datetime
-    from ue.inbox.gmail import sync_gmail_inbox, sync_gmail_sent
-    from ue.inbox.calendar import sync_calendar
-    from ue.activity.git import sync_git_commits
-    from ue.config import set_last_sync
+# Import and register commands from modules
+from ue.commands.sync import (
+    sync, dashboard, dashboard_short, inbox, calendar, activity, add_repo
+)
+from ue.commands.task import task, done
+from ue.commands.block import block, did
+from ue.commands.routines import am, pm, review, status, focus
+from ue.commands.log import log, mark
+from ue.commands.demo import demo_setup, demo_reset
 
-    if not quiet:
-        console.print("[bold]Syncing...[/bold]\n")
-
-    try:
-        result = sync_gmail_inbox(days=days)
-        if not quiet:
-            console.print(f"  Gmail inbox: {result['fetched']} emails")
-    except Exception as e:
-        if not quiet:
-            console.print(f"  [red]Gmail inbox error: {e}[/red]")
-
-    try:
-        result = sync_gmail_sent(days=days)
-        skipped = result.get('skipped', 0)
-        if not quiet:
-            msg = f"  Gmail sent: {result['logged']} emails logged"
-            if skipped:
-                msg += f" ({skipped} already synced)"
-            console.print(msg)
-    except Exception as e:
-        if not quiet:
-            console.print(f"  [red]Gmail sent error: {e}[/red]")
-
-    try:
-        result = sync_calendar(days_ahead=days)
-        if not quiet:
-            console.print(f"  Calendar: {result['fetched']} events")
-    except Exception as e:
-        if not quiet:
-            console.print(f"  [red]Calendar error: {e}[/red]")
-
-    try:
-        result = sync_git_commits(since_days=days)
-        skipped = result.get('skipped', 0)
-        if not quiet:
-            msg = f"  Git: {result['logged']} commits from {result['repos_scanned']} repos"
-            if skipped:
-                msg += f" ({skipped} already synced)"
-            console.print(msg)
-    except Exception as e:
-        if not quiet:
-            console.print(f"  [red]Git error: {e}[/red]")
-
-    set_last_sync(datetime.now().isoformat())
-
-    if not quiet:
-        console.print("\n[green]Sync complete![/green]")
-
-
-def auto_sync_if_stale():
-    """Auto-sync if data is stale. Returns True if sync was run."""
-    from ue.config import is_sync_stale
-
-    if is_sync_stale():
-        console.print("[dim]Auto-syncing...[/dim]")
-        run_sync(quiet=True)
-        console.print("[dim]Sync complete.[/dim]\n")
-        return True
-    return False
-
-
-@cli.command()
-@click.option("--days", default=7, help="Days of history to sync")
-def sync(days):
-    """Sync data from Gmail, Calendar, and git."""
-    run_sync(days=days, quiet=False)
-
-
-@cli.command()
-def dashboard():
-    """Show the main dashboard."""
-    auto_sync_if_stale()
-    from ue.dashboard import show_dashboard
-    show_dashboard()
-
-
-# Alias 'd' for dashboard
-@cli.command("d")
-def dashboard_short():
-    """Show the main dashboard (shortcut)."""
-    auto_sync_if_stale()
-    from ue.dashboard import show_dashboard
-    show_dashboard()
-
-
-@cli.command()
-@click.option("--limit", "-n", default=20, help="Number of items to show")
-def inbox(limit):
-    """Show email inbox."""
-    from ue.dashboard import show_inbox
-    show_inbox(source="gmail", limit=limit)
-
-
-@cli.command()
-@click.option("--days", "-d", default=7, help="Days ahead to show")
-def calendar(days):
-    """Show upcoming calendar events."""
-    from ue.dashboard import show_calendar
-    show_calendar(days=days)
-
-
-@cli.command()
-@click.option("--days", "-d", default=7, help="Days of history")
-@click.option("--type", "-t", "activity_type", help="Filter by type")
-def activity(days, activity_type):
-    """Show activity log."""
-    from ue.dashboard import show_activity
-    show_activity(days=days, activity_type=activity_type)
-
-
-@cli.group()
-def log():
-    """Log activities manually."""
-    pass
-
-
-@log.command("application")
-@click.argument("company")
-@click.option("--role", "-r", help="Role/position applied for")
-@click.option("--notes", "-n", help="Additional notes")
-def log_application_cmd(company, role, notes):
-    """Log a job application."""
-    from ue.activity.manual import log_application
-    log_application(company=company, role=role, notes=notes)
-    console.print(f"[green]Logged application to {company}[/green]")
-
-
-@log.command("win")
-@click.argument("description")
-@click.option("--workstream", "-w", help="Workstream (ai-research, terrasol, blog, consulting)")
-@click.option("--notes", "-n", help="Additional notes")
-def log_win_cmd(description, workstream, notes):
-    """Log a win or accomplishment."""
-    from ue.activity.manual import log_win
-    log_win(description=description, workstream=workstream, notes=notes)
-    console.print(f"[green]Logged win: {description}[/green]")
-
-
-@cli.group()
-def mark():
-    """Mark inbox items."""
-    pass
-
-
-@mark.command("respond")
-@click.argument("item_id")
-def mark_needs_response(item_id):
-    """Mark an item as needing response."""
-    from ue.db import get_db
-
-    db = get_db()
-    # Handle both numeric index and full ID
-    if item_id.isdigit():
-        # Get by row number
-        items = db.execute(
-            "SELECT id FROM inbox_items ORDER BY timestamp DESC LIMIT ? OFFSET ?",
-            (1, int(item_id) - 1)
-        ).fetchall()
-        if items:
-            item_id = items[0]["id"]
-        else:
-            console.print("[red]Item not found[/red]")
-            return
-
-    db.execute("UPDATE inbox_items SET needs_response = 1 WHERE id = ?", (item_id,))
-    db.commit()
-    db.close()
-    console.print(f"[green]Marked as needing response[/green]")
-
-
-@mark.command("done")
-@click.argument("item_id")
-def mark_responded(item_id):
-    """Mark an item as responded to."""
-    from ue.db import get_db
-
-    db = get_db()
-    if item_id.isdigit():
-        items = db.execute(
-            "SELECT id FROM inbox_items ORDER BY timestamp DESC LIMIT ? OFFSET ?",
-            (1, int(item_id) - 1)
-        ).fetchall()
-        if items:
-            item_id = items[0]["id"]
-        else:
-            console.print("[red]Item not found[/red]")
-            return
-
-    db.execute(
-        "UPDATE inbox_items SET needs_response = 0, responded = 1 WHERE id = ?",
-        (item_id,)
-    )
-    db.commit()
-    db.close()
-    console.print(f"[green]Marked as done[/green]")
-
-
-@mark.command("workstream")
-@click.argument("item_id")
-@click.argument("workstream")
-def mark_workstream(item_id, workstream):
-    """Assign a workstream to an item."""
-    from ue.db import get_db
-
-    valid = ["ai-research", "terrasol", "blog", "consulting"]
-    if workstream not in valid:
-        console.print(f"[red]Invalid workstream. Choose from: {', '.join(valid)}[/red]")
-        return
-
-    db = get_db()
-    if item_id.isdigit():
-        items = db.execute(
-            "SELECT id FROM inbox_items ORDER BY timestamp DESC LIMIT ? OFFSET ?",
-            (1, int(item_id) - 1)
-        ).fetchall()
-        if items:
-            item_id = items[0]["id"]
-        else:
-            console.print("[red]Item not found[/red]")
-            return
-
-    db.execute("UPDATE inbox_items SET workstream = ? WHERE id = ?", (workstream, item_id))
-    db.commit()
-    db.close()
-    console.print(f"[green]Assigned to {workstream}[/green]")
-
-
-@cli.command()
-@click.argument("repo_path")
-def add_repo(repo_path):
-    """Add a git repository to track."""
-    from pathlib import Path
-    from ue.config import load_config, save_config
-
-    path = Path(repo_path).expanduser().resolve()
-    if not (path / ".git").exists():
-        console.print(f"[red]{path} is not a git repository[/red]")
-        return
-
-    config = load_config()
-    repos = config.get("git_repos", [])
-    if str(path) not in repos:
-        repos.append(str(path))
-        config["git_repos"] = repos
-        save_config(config)
-        console.print(f"[green]Added {path}[/green]")
-    else:
-        console.print(f"[yellow]Already tracking {path}[/yellow]")
-
+# Sync/display commands
+cli.add_command(sync)
+cli.add_command(dashboard)
+cli.add_command(dashboard_short)
+cli.add_command(inbox)
+cli.add_command(calendar)
+cli.add_command(activity)
+cli.add_command(add_repo)
 
 # Task commands
-@cli.group()
-def task():
-    """Manage time-sensitive tasks."""
-    pass
-
-
-def parse_due_date(due: str) -> str:
-    """Parse natural language dates into YYYY-MM-DD format."""
-    from datetime import datetime, timedelta
-
-    if not due:
-        return None
-
-    due_lower = due.lower().strip()
-    today = datetime.now().date()
-
-    day_names = ["monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"]
-    day_abbrevs = ["mon", "tue", "wed", "thu", "fri", "sat", "sun"]
-
-    if due_lower == "today":
-        return today.isoformat()
-    elif due_lower == "tomorrow":
-        return (today + timedelta(days=1)).isoformat()
-    elif due_lower.startswith("next "):
-        # Handle "next monday", "next wed", etc.
-        day_part = due_lower[5:]  # Remove "next "
-        if day_part in day_abbrevs:
-            target_day = day_abbrevs.index(day_part)
-        elif day_part in day_names:
-            target_day = day_names.index(day_part)
-        else:
-            return due  # Can't parse, return as-is
-        # "next X" means the X in the following week
-        days_ahead = target_day - today.weekday()
-        if days_ahead <= 0:
-            days_ahead += 7
-        days_ahead += 7  # Add a week for "next"
-        return (today + timedelta(days=days_ahead)).isoformat()
-    elif due_lower in day_names or due_lower in day_abbrevs:
-        # Find next occurrence of that day
-        if due_lower in day_abbrevs:
-            target_day = day_abbrevs.index(due_lower)
-        else:
-            target_day = day_names.index(due_lower)
-        days_ahead = target_day - today.weekday()
-        if days_ahead <= 0:
-            days_ahead += 7
-        return (today + timedelta(days=days_ahead)).isoformat()
-    else:
-        # Assume YYYY-MM-DD format
-        return due
-
-
-@task.command("add")
-@click.argument("title", required=False, default=None)
-@click.option("--due", "-d", help="Due date (YYYY-MM-DD or 'wed', 'friday', etc.)")
-@click.option("--workstream", "-w", help="Workstream")
-@click.option("--priority", "-p", type=click.Choice(["low", "normal", "high"]), default=None)
-@click.option("--notes", "-n", help="Additional notes")
-def task_add(title, due, workstream, priority, notes):
-    """Add a task with a deadline. Run without args for interactive mode."""
-    from rich.prompt import Prompt
-    from ue.db import add_task
-    from ue.config import load_config
-
-    # Interactive mode if no title provided
-    if title is None:
-        console.print("\n[bold]Add a task[/bold]\n")
-
-        # 1. Get task title
-        title = Prompt.ask("  Task name")
-        if not title.strip():
-            console.print("[red]Task name required[/red]")
-            return
-
-        # 2. Select priority
-        priorities = ["low", "normal", "high"]
-        console.print("\n  [bold]Priority:[/bold]")
-        for i, p in enumerate(priorities, 1):
-            marker = ""
-            if p == "high":
-                marker = " [red](urgent)[/red]"
-            elif p == "normal":
-                marker = " [dim](default)[/dim]"
-            console.print(f"    {i}. {p}{marker}")
-
-        priority_choice = Prompt.ask("\n  Select priority", default="2")
-        try:
-            idx = int(priority_choice) - 1
-            if 0 <= idx < len(priorities):
-                priority = priorities[idx]
-            else:
-                priority = "normal"
-        except ValueError:
-            priority = "normal"
-
-        # 3. Select workstream
-        config = load_config()
-        workstreams = list(config.get("workstreams", {}).keys())
-
-        if workstreams:
-            console.print("\n  [bold]Workstream:[/bold]")
-            console.print("    0. [dim](none)[/dim]")
-            for i, ws in enumerate(workstreams, 1):
-                console.print(f"    {i}. {ws}")
-
-            ws_choice = Prompt.ask("\n  Select workstream", default="0")
-            try:
-                idx = int(ws_choice)
-                if idx == 0:
-                    workstream = None
-                elif 1 <= idx <= len(workstreams):
-                    workstream = workstreams[idx - 1]
-                else:
-                    workstream = None
-            except ValueError:
-                workstream = None
-
-        # 4. Due date (optional)
-        due = Prompt.ask("\n  Due date (today, tomorrow, wed, 2024-01-15, or skip)", default="")
-        console.print()
-
-    # Parse due date
-    due_date = parse_due_date(due) if due else None
-
-    # Default priority if not set
-    if priority is None:
-        priority = "normal"
-
-    task_id = add_task(
-        title=title,
-        due_date=due_date,
-        workstream=workstream,
-        priority=priority,
-        notes=notes,
-    )
-
-    if due_date:
-        console.print(f"[green]Added task #{task_id}: {title} (due {due_date})[/green]")
-    else:
-        console.print(f"[green]Added task #{task_id}: {title}[/green]")
-
-
-@task.command("list")
-@click.option("--all", "show_all", is_flag=True, help="Include completed tasks")
-def task_list(show_all):
-    """List pending tasks."""
-    from datetime import datetime
-    from rich.table import Table
-    from ue.db import get_tasks, get_overdue_tasks
-
-    tasks = get_tasks(status="pending")
-    overdue = {t["id"] for t in get_overdue_tasks()}
-    today = datetime.now().strftime("%Y-%m-%d")
-
-    if not tasks:
-        console.print("[dim]No pending tasks[/dim]")
-        return
-
-    table = Table(show_header=True, header_style="bold")
-    table.add_column("ID", width=4)
-    table.add_column("Task", width=40)
-    table.add_column("Due", width=12)
-    table.add_column("Pri", width=6)
-    table.add_column("Stream", width=12)
-
-    for t in tasks:
-        due_str = t["due_date"] or "-"
-        if t["id"] in overdue:
-            due_str = f"[red]{due_str} ![/red]"
-        elif t["due_date"] == today:
-            due_str = f"[yellow]{due_str}[/yellow]"
-
-        pri_str = t["priority"]
-        if t["priority"] == "high":
-            pri_str = f"[red]{pri_str}[/red]"
-
-        table.add_row(
-            str(t["id"]),
-            t["title"][:40],
-            due_str,
-            pri_str,
-            t["workstream"] or "-",
-        )
-
-    console.print(table)
-
-
-@task.command("done")
-@click.argument("task_id", type=int)
-def task_done(task_id):
-    """Mark a task as complete."""
-    from ue.db import complete_task
-    complete_task(task_id)
-    console.print(f"[green]Task #{task_id} completed![/green]")
-
-
-@task.command("cancel")
-@click.argument("task_id", type=int)
-def task_cancel(task_id):
-    """Cancel a task."""
-    from ue.db import cancel_task
-    cancel_task(task_id)
-    console.print(f"[yellow]Task #{task_id} cancelled[/yellow]")
-
-
-@cli.command()
-def did():
-    """Interactive block completion - pick a block you just did."""
-    from datetime import datetime
-    from rich.prompt import Prompt
-    from ue.db import get_block_targets, get_block_completions, log_block_completion, get_week_block_summary
-
-    targets = get_block_targets()
-
-    if not targets:
-        console.print("[dim]No blocks configured. Use 'ue block target <name> <weekly_count>'[/dim]")
-        return
-
-    today = get_effective_date().isoformat()
-    today_completions = {c["block_name"]: c for c in get_block_completions(since=today)}
-
-    console.print("\n[bold]Blocks:[/bold]\n")
-    for i, t in enumerate(targets, 1):
-        name = t["block_name"]
-        weekly = t["weekly_target"]
-        summary = get_week_block_summary(name)
-
-        # Status indicator
-        if name in today_completions:
-            status = today_completions[name]["status"]
-            if status == "completed":
-                status_str = " [green]✓ done today[/green]"
-            elif status == "skipped":
-                status_str = " [yellow]skipped today[/yellow]"
-            else:
-                status_str = " [blue]partial today[/blue]"
-        else:
-            status_str = ""
-
-        # Progress
-        if weekly == 0:
-            progress = "[dim](daily)[/dim]"
-        else:
-            progress = f"[dim]({summary['completed']}/{weekly} this week)[/dim]"
-
-        console.print(f"  {i}. {name} {progress}{status_str}")
-
-    console.print()
-    choice = Prompt.ask(
-        "Which did you do? (number or 'q' to quit)",
-        default="q"
-    )
-
-    if choice.lower() == 'q':
-        return
-
-    try:
-        idx = int(choice) - 1
-        if 0 <= idx < len(targets):
-            block = targets[idx]
-            log_block_completion(block["block_name"], today, "completed")
-            console.print(f"\n[green]Logged: {block['block_name']} ✓[/green]")
-        else:
-            console.print("[red]Invalid number[/red]")
-    except ValueError:
-        console.print("[red]Invalid input[/red]")
-
-
-@cli.command()
-def done():
-    """Interactive task completion - pick a task to mark done."""
-    from datetime import datetime
-    from rich.prompt import Prompt
-    from ue.db import get_tasks, complete_task, get_overdue_tasks
-
-    tasks = get_tasks(status="pending")
-
-    if not tasks:
-        console.print("[dim]No pending tasks[/dim]")
-        return
-
-    overdue = {t["id"] for t in get_overdue_tasks()}
-    today = get_effective_date().isoformat()
-
-    console.print("\n[bold]Pending tasks:[/bold]\n")
-    for i, t in enumerate(tasks, 1):
-        due_str = ""
-        if t["due_date"]:
-            if t["id"] in overdue:
-                due_str = f" [red](overdue: {t['due_date']})[/red]"
-            elif t["due_date"] == today:
-                due_str = f" [yellow](due today)[/yellow]"
-            else:
-                due_str = f" [dim](due {t['due_date']})[/dim]"
-
-        console.print(f"  {i}. {t['title']}{due_str}")
-
-    console.print()
-    choice = Prompt.ask(
-        "Which task is done? (number or 'q' to quit)",
-        default="q"
-    )
-
-    if choice.lower() == 'q':
-        return
-
-    try:
-        idx = int(choice) - 1
-        if 0 <= idx < len(tasks):
-            task = tasks[idx]
-            complete_task(task["id"])
-            console.print(f"\n[green]Completed: {task['title']}[/green]")
-        else:
-            console.print("[red]Invalid number[/red]")
-    except ValueError:
-        console.print("[red]Invalid input[/red]")
-
-
-# Block tracking commands
-@cli.group()
-def block():
-    """Track time block completions."""
-    pass
-
-
-@block.command("done")
-@click.argument("name")
-@click.option("--date", "-d", default=None, help="Date (YYYY-MM-DD), defaults to today")
-@click.option("--notes", "-n", help="Optional notes")
-@click.option("--minutes", "-m", type=int, help="Duration in minutes")
-def block_done(name, date, notes, minutes):
-    """Mark a block as completed."""
-    from datetime import datetime
-    from ue.db import log_block_completion
-
-    if date is None:
-        date = get_effective_date().isoformat()
-
-    log_block_completion(
-        block_name=name,
-        date=date,
-        status="completed",
-        notes=notes,
-        duration_minutes=minutes,
-    )
-    console.print(f"[green]Marked '{name}' as completed for {date}[/green]")
-
-
-@block.command("skip")
-@click.argument("name")
-@click.option("--reason", "-r", required=True, help="Why was it skipped?")
-@click.option("--date", "-d", default=None, help="Date (YYYY-MM-DD), defaults to today")
-def block_skip(name, reason, date):
-    """Mark a block as skipped with a reason."""
-    from datetime import datetime
-    from ue.db import log_block_completion
-
-    if date is None:
-        date = get_effective_date().isoformat()
-
-    log_block_completion(
-        block_name=name,
-        date=date,
-        status="skipped",
-        reason=reason,
-    )
-    console.print(f"[yellow]Marked '{name}' as skipped for {date}[/yellow]")
-
-
-@block.command("partial")
-@click.argument("name")
-@click.option("--reason", "-r", help="What happened?")
-@click.option("--date", "-d", default=None, help="Date (YYYY-MM-DD), defaults to today")
-@click.option("--minutes", "-m", type=int, help="Duration in minutes")
-def block_partial(name, reason, date, minutes):
-    """Mark a block as partially completed."""
-    from datetime import datetime
-    from ue.db import log_block_completion
-
-    if date is None:
-        date = get_effective_date().isoformat()
-
-    log_block_completion(
-        block_name=name,
-        date=date,
-        status="partial",
-        reason=reason,
-        duration_minutes=minutes,
-    )
-    console.print(f"[blue]Marked '{name}' as partial for {date}[/blue]")
-
-
-@block.command("target")
-@click.argument("name")
-@click.argument("weekly_count", type=int)
-@click.option("--workstream", "-w", help="Associated workstream")
-def block_target(name, weekly_count, workstream):
-    """Set weekly target for a block (0 = daily)."""
-    from ue.db import set_block_target
-
-    set_block_target(block_name=name, weekly_target=weekly_count, workstream=workstream)
-    if weekly_count == 0:
-        console.print(f"[green]Set '{name}' as a daily block[/green]")
-    else:
-        console.print(f"[green]Set '{name}' target: {weekly_count}x per week[/green]")
-
-
-@block.command("list")
-def block_list():
-    """Show today's blocks and weekly progress."""
-    from datetime import datetime, timedelta
-    from rich.table import Table
-    from ue.db import get_block_targets, get_block_completions, get_week_block_summary
-
-    targets = get_block_targets()
-    today = get_effective_date()
-    today_str = today.isoformat()
-    week_start = (today - timedelta(days=today.weekday())).isoformat()
-
-    if not targets:
-        console.print("[dim]No block targets set. Use 'ue block target <name> <weekly_count>'[/dim]")
-        return
-
-    table = Table(title=f"Blocks - Week of {week_start}", show_header=True, header_style="bold")
-    table.add_column("Block", width=20)
-    table.add_column("Target", width=10)
-    table.add_column("Done", width=8)
-    table.add_column("Today", width=10)
-    table.add_column("Stream", width=12)
-
-    # Get today's completions
-    today_completions = {c["block_name"]: c for c in get_block_completions(since=today_str)}
-
-    for target in targets:
-        name = target["block_name"]
-        weekly = target["weekly_target"]
-        summary = get_week_block_summary(name)
-
-        target_str = "daily" if weekly == 0 else f"{weekly}x/wk"
-        done_str = str(summary["completed"])
-
-        # Color coding
-        if weekly > 0:
-            if summary["completed"] >= weekly:
-                done_str = f"[green]{done_str}[/green]"
-            elif summary["completed"] >= weekly // 2:
-                done_str = f"[yellow]{done_str}[/yellow]"
-            else:
-                done_str = f"[red]{done_str}[/red]"
-
-        # Today's status
-        today_status = "-"
-        if name in today_completions:
-            status = today_completions[name]["status"]
-            if status == "completed":
-                today_status = "[green]done[/green]"
-            elif status == "skipped":
-                today_status = "[red]skipped[/red]"
-            elif status == "partial":
-                today_status = "[yellow]partial[/yellow]"
-
-        table.add_row(
-            name,
-            target_str,
-            done_str,
-            today_status,
-            target.get("workstream") or "-",
-        )
-
-    console.print(table)
-
-
-def get_at_risk_blocks():
-    """Calculate which blocks are at risk or behind pace this week."""
-    from datetime import datetime, timedelta
-    from ue.db import get_block_targets, get_week_block_summary, get_block_completions
-    from ue.config import load_config
-
-    today = get_effective_date()
-    today_str = today.isoformat()
-    # Days elapsed and left in week (0 = Monday, 6 = Sunday)
-    day_of_week = today.weekday()
-    days_elapsed = day_of_week + 1  # Mon=1, Tue=2, etc.
-    days_left = 6 - day_of_week  # Including today
-
-    # Get workstream priorities
-    config = load_config()
-    workstreams = config.get("workstreams", {})
-
-    def get_ws_priority(workstream):
-        if workstream and workstream in workstreams:
-            return workstreams[workstream].get("priority", "low")
-        return None  # No workstream
-
-    targets = get_block_targets()
-    today_completions = {c["block_name"]: c for c in get_block_completions(since=today_str)}
-    at_risk = []
-
-    for target in targets:
-        name = target["block_name"]
-        weekly_target = target["weekly_target"]
-        workstream = target.get("workstream")
-        ws_priority = get_ws_priority(workstream)
-        summary = get_week_block_summary(name)
-        completed = summary["completed"]
-
-        if weekly_target == 0:
-            # Daily block - flag if not done today
-            if name not in today_completions:
-                at_risk.append({
-                    "name": name,
-                    "target": "daily",
-                    "completed": completed,
-                    "remaining": 1,
-                    "days_left": days_left,
-                    "workstream": workstream,
-                    "status": "daily_pending"
-                })
-            continue
-
-        remaining = weekly_target - completed
-
-        if remaining <= 0:
-            # Already hit target
-            continue
-
-        slack = days_left - remaining  # How many days we can skip and still hit target
-
-        if remaining > days_left:
-            # Can't possibly hit target
-            at_risk.append({
-                "name": name,
-                "target": weekly_target,
-                "completed": completed,
-                "remaining": remaining,
-                "days_left": days_left,
-                "workstream": workstream,
-                "status": "impossible"
-            })
-        elif slack <= 1:
-            # At risk - only 0 or 1 day of slack
-            at_risk.append({
-                "name": name,
-                "target": weekly_target,
-                "completed": completed,
-                "remaining": remaining,
-                "days_left": days_left,
-                "workstream": workstream,
-                "status": "at_risk"
-            })
-        elif ws_priority in ("high", "mid") and completed < weekly_target / 2:
-            # High/mid priority and under halfway - try to do
-            at_risk.append({
-                "name": name,
-                "target": weekly_target,
-                "completed": completed,
-                "remaining": remaining,
-                "days_left": days_left,
-                "workstream": workstream,
-                "status": "try_to_do"
-            })
-        elif ws_priority == "low" and slack <= 2:
-            # Low priority - only flag when slack is tight (2 days or less)
-            at_risk.append({
-                "name": name,
-                "target": weekly_target,
-                "completed": completed,
-                "remaining": remaining,
-                "days_left": days_left,
-                "workstream": workstream,
-                "status": "try_to_do"
-            })
-        # No workstream and not urgent = don't show
-
-    return at_risk
-
-
-@cli.command()
-def am():
-    """Morning standup - start your day."""
-    auto_sync_if_stale()
-    from datetime import datetime, timedelta
-    from rich.panel import Panel
-    from ue.db import get_inbox_items, get_block_targets, get_block_completions
-    from ue.db import get_upcoming_tasks, get_overdue_tasks
-    from ue.inbox.calendar import get_upcoming_events
-
-    today = get_effective_date()
-    today_str = today.isoformat()
-    day_name = today.strftime("%A")
-    date_str = today.strftime("%B %d, %Y")
-
-    console.print()
-    print_logo()
-    console.print(Panel(f"[bold]Good morning![/bold] {day_name}, {date_str}", style="blue"))
-
-    # Overdue and upcoming tasks (DEADLINES)
-    overdue = get_overdue_tasks()
-    upcoming = get_upcoming_tasks(days=7)
-
-    if overdue:
-        console.print("\n[bold red]OVERDUE[/bold red]")
-        for t in overdue:
-            console.print(f"  [red]  {t['title']} (was due {t['due_date']})[/red]")
-
-    # Show tasks due soon (not overdue)
-    due_soon = [t for t in upcoming if t["due_date"] >= today_str]
-    if due_soon:
-        console.print("\n[bold yellow]DEADLINES[/bold yellow]")
-        for t in due_soon[:5]:
-            days_until = (datetime.strptime(t["due_date"], "%Y-%m-%d").date() - today).days
-            if days_until == 0:
-                due_str = "[yellow]today[/yellow]"
-            elif days_until == 1:
-                due_str = "[yellow]tomorrow[/yellow]"
-            else:
-                due_str = f"in {days_until} days"
-            console.print(f"    {t['title']} ({due_str})")
-
-    # At-risk and behind-pace blocks
-    at_risk = get_at_risk_blocks()
-    if at_risk:
-        from ue.config import load_config
-
-        # Sort by urgency: impossible > at_risk > try_to_do > daily_pending
-        # Then by workstream priority within each tier
-        config = load_config()
-        workstreams = config.get("workstreams", {})
-        priority_order = {"high": 0, "mid": 1, "low": 2}
-        status_order = {"impossible": 0, "at_risk": 1, "try_to_do": 2, "daily_pending": 3}
-
-        def sort_key(block):
-            ws = block.get("workstream")
-            ws_pri = priority_order.get(workstreams.get(ws, {}).get("priority", "low"), 2) if ws else 3
-            return (status_order.get(block["status"], 4), ws_pri)
-
-        sorted_blocks = sorted(at_risk, key=sort_key)
-
-        # Show all critical (impossible/at_risk), but limit try_to_do to keep total at 3 max
-        critical = [b for b in sorted_blocks if b["status"] in ("impossible", "at_risk")]
-        others = [b for b in sorted_blocks if b["status"] not in ("impossible", "at_risk")]
-        max_others = max(0, 3 - len(critical))
-        to_show = critical + others[:max_others]
-
-        console.print("\n[bold yellow]BLOCKS[/bold yellow]")
-        for block in to_show:
-            if block["status"] == "impossible":
-                console.print(f"  [red]{block['name']}: {block['completed']}/{block['target']} - can't hit target[/red]")
-            elif block["status"] == "at_risk":
-                console.print(f"  [red]{block['name']}: {block['completed']}/{block['target']} - at risk[/red]")
-            elif block["status"] == "try_to_do":
-                console.print(f"  [yellow]{block['name']}: {block['completed']}/{block['target']} - try to do[/yellow]")
-            elif block["status"] == "daily_pending":
-                console.print(f"  [yellow]{block['name']}: not done today[/yellow]")
-    else:
-        console.print("\n[green]All blocks on track[/green]")
-
-    # Today's calendar
-    console.print("\n[bold]TODAY'S CALENDAR[/bold]")
-    try:
-        events = get_upcoming_events(hours=12)
-        if events:
-            for event in events[:8]:
-                start = event.get("start", {})
-                time_str = ""
-                if "dateTime" in start:
-                    dt = datetime.fromisoformat(start["dateTime"].replace("Z", "+00:00"))
-                    time_str = dt.strftime("%I:%M %p")
-                summary = event.get("summary", "(no title)")
-                console.print(f"  {time_str:>8}  {summary}")
-        else:
-            console.print("  [dim]No events today[/dim]")
-    except Exception as e:
-        console.print(f"  [dim]Could not fetch calendar: {e}[/dim]")
-
-    # Inbox needing response
-    needs_response = get_inbox_items(needs_response=True, limit=5)
-    if needs_response:
-        console.print(f"\n[bold yellow]NEEDS RESPONSE ({len(needs_response)})[/bold yellow]")
-        for item in needs_response[:5]:
-            age_days = (today - datetime.fromisoformat(item["timestamp"].replace("Z", ""))).days
-            age_str = f"{age_days}d" if age_days > 0 else "today"
-            console.print(f"  [{age_str}] {item['source']}: {(item['subject'] or '')[:50]}")
-
-    # Recent commits by repo
-    from ue.db import get_activity
-    import json as json_mod
-    week_start = today - timedelta(days=today.weekday())
-    week_activity = get_activity(activity_type="commit", since=week_start.isoformat())
-    if week_activity:
-        console.print("\n[bold]COMMITS THIS WEEK[/bold]")
-        # Group by repo
-        by_repo = {}
-        for act in week_activity:
-            if act.get("metadata"):
-                try:
-                    meta = json_mod.loads(act["metadata"])
-                    repo = meta.get("repo", "unknown")
-                    by_repo[repo] = by_repo.get(repo, 0) + 1
-                except Exception:
-                    pass
-        for repo, count in sorted(by_repo.items(), key=lambda x: -x[1]):
-            console.print(f"  {repo}: {count} commit{'s' if count != 1 else ''}")
-
-    # Suggested focus
-    console.print("\n[bold]SUGGESTED FOCUS[/bold]")
-    if at_risk:
-        from ue.config import load_config
-        from ue.db import get_block_targets
-
-        # Get workstream priorities
-        config = load_config()
-        workstreams = config.get("workstreams", {})
-        priority_order = {"high": 0, "mid": 1, "low": 2}
-
-        # Get block -> workstream mapping
-        block_targets = {t["block_name"]: t.get("workstream") for t in get_block_targets()}
-
-        def get_priority(block):
-            ws = block_targets.get(block["name"])
-            if ws and ws in workstreams:
-                return priority_order.get(workstreams[ws].get("priority", "low"), 2)
-            return 3  # No workstream = lowest priority
-
-        # Priority order: impossible > at_risk > try_to_do > daily_pending
-        # Within each tier, sort by workstream priority
-        critical = sorted([b for b in at_risk if b["status"] in ("impossible", "at_risk")], key=get_priority)
-        try_to_do = sorted([b for b in at_risk if b["status"] == "try_to_do"], key=get_priority)
-        daily = sorted([b for b in at_risk if b["status"] == "daily_pending"], key=get_priority)
-
-        if critical:
-            console.print(f"  [red]Priority: {critical[0]['name']} ({critical[0]['completed']}/{critical[0]['target']} - at risk)[/red]")
-        elif try_to_do:
-            console.print(f"  [yellow]Try to do: {try_to_do[0]['name']} ({try_to_do[0]['completed']}/{try_to_do[0]['target']})[/yellow]")
-        elif daily:
-            console.print(f"  [yellow]Today: {', '.join(b['name'] for b in daily)}[/yellow]")
-    else:
-        console.print("  [green]All blocks on track - nice![/green]")
-
-    console.print()
-
-
-@cli.command()
-def pm():
-    """Evening review - end your day."""
-    from datetime import datetime
-    from rich.prompt import Prompt
-    from rich.panel import Panel
-    from ue.db import get_block_targets, get_block_completions, log_block_completion, get_activity
-    from ue.activity.manual import log_win
-
-    today = get_effective_date()
-    today_str = today.isoformat()
-
-    console.print()
-    console.print(Panel("[bold]Evening Review[/bold]", style="magenta"))
-
-    # Show what was accomplished today
-    today_activity = get_activity(since=today_str + "T00:00:00")
-    if today_activity:
-        console.print("\n[bold]TODAY'S ACTIVITY[/bold]")
-        by_type = {}
-        for act in today_activity:
-            t = act["activity_type"]
-            by_type[t] = by_type.get(t, 0) + 1
-        for t, count in by_type.items():
-            console.print(f"  {t}: {count}")
-
-    # Walk through blocks
-    console.print("\n[bold]BLOCK CHECK-IN[/bold]")
-    targets = get_block_targets()
-    today_completions = {c["block_name"]: c for c in get_block_completions(since=today_str)}
-
-    for target in targets:
-        name = target["block_name"]
-        weekly = target["weekly_target"]
-
-        # Skip 1x/week blocks if already done this week
-        if weekly > 0 and weekly == 1:
-            from ue.db import get_week_block_summary
-            summary = get_week_block_summary(name)
-            if summary["completed"] >= weekly:
-                continue
-
-        if name in today_completions:
-            status = today_completions[name]["status"]
-            console.print(f"  [dim]{name}: {status}[/dim]")
-            continue
-
-        # Only ask about daily blocks or blocks not yet hit this week
-        if weekly == 0 or True:  # Ask about all for now
-            response = Prompt.ask(
-                f"  {name}",
-                choices=["done", "skip", "partial", "n/a"],
-                default="n/a"
-            )
-
-            if response == "done":
-                log_block_completion(name, today_str, "completed")
-                console.print(f"    [green]done[/green]")
-            elif response == "skip":
-                reason = Prompt.ask("    Why?", default="")
-                log_block_completion(name, today_str, "skipped", reason=reason if reason else None)
-                console.print(f"    [yellow]skipped[/yellow]")
-            elif response == "partial":
-                reason = Prompt.ask("    What happened?", default="")
-                log_block_completion(name, today_str, "partial", reason=reason if reason else None)
-                console.print(f"    [blue]partial[/blue]")
-            # n/a = don't log anything
-
-    # Capture any wins
-    console.print("\n[bold]WINS[/bold]")
-    win = Prompt.ask("  Any wins today? (or enter to skip)", default="")
-    if win:
-        log_win(win)
-        console.print(f"    [green]logged![/green]")
-
-    # Preview tomorrow
-    at_risk = get_at_risk_blocks()
-    if at_risk:
-        console.print("\n[bold]TOMORROW'S PRIORITY[/bold]")
-        for block in at_risk[:2]:
-            console.print(f"  [yellow]{block['name']}: {block['remaining']} remaining, {block['days_left']} days left[/yellow]")
-
-    console.print("\n[magenta]Good night![/magenta]\n")
-
-
-# Keep old review as alias
-@cli.command()
-def review():
-    """Daily review (alias for 'pm')."""
-    from click import Context
-    ctx = Context(pm)
-    ctx.invoke(pm)
-
-
-@cli.command()
-@click.option("--copy", "copy_mode", is_flag=True, help="Print context for manual copy/paste instead of calling API")
-def focus(copy_mode):
-    """Get AI-powered recommendation for what to focus on now."""
-    from ue.focus import get_focus, print_context
-
-    if copy_mode:
-        console.print(print_context())
-        console.print("\n[dim]Copy the above and paste into claude.ai[/dim]")
-        return
-
-    try:
-        result = get_focus()
-        console.print()
-        console.print(result)
-        console.print()
-    except ValueError as e:
-        console.print(f"[red]{e}[/red]")
-    except Exception as e:
-        console.print(f"[red]Error calling API: {e}[/red]")
-        console.print("[dim]Try --copy mode to use claude.ai instead[/dim]")
-
-
-@cli.command()
-def status():
-    """Show week-to-date status with completed and pending tasks."""
-    auto_sync_if_stale()
-    from datetime import datetime, timedelta
-    from rich.panel import Panel
-    from rich.table import Table
-    from ue.db import (
-        get_tasks, get_tasks_completed_since, get_overdue_tasks,
-        get_block_targets, get_block_completions, get_week_block_summary
-    )
-
-    today = get_effective_date()
-    week_start = today - timedelta(days=today.weekday())
-    week_start_str = week_start.isoformat()
-    day_name = today.strftime("%A")
-    day_of_week = today.weekday()  # 0=Mon, 6=Sun
-    days_elapsed = day_of_week + 1  # Mon=1, Sun=7
-
-    console.print()
-    console.print(Panel(
-        f"[bold]Week Status[/bold] - {week_start.strftime('%b %d')} to {today.strftime('%b %d')} ({day_name})",
-        style="cyan"
-    ))
-
-    # === BLOCKS SECTION ===
-    targets = get_block_targets()
-    if targets:
-        # Get all block completions this week
-        week_completions = get_block_completions(since=week_start_str)
-
-        # Count total block completions this week
-        total_block_completions = sum(1 for c in week_completions if c["status"] == "completed")
-
-        console.print(f"\n[bold]Blocks[/bold]")
-
-        for target in targets:
-            name = target["block_name"]
-            weekly = target["weekly_target"]
-            summary = get_week_block_summary(name)
-            completed = summary["completed"]
-
-            if weekly == 0:
-                # Daily block - show completed/days elapsed
-                target_str = f"{completed}/{days_elapsed} days"
-                if completed >= days_elapsed:
-                    console.print(f"  [green]{name}: {target_str}[/green]")
-                elif completed >= days_elapsed - 1:
-                    console.print(f"  [yellow]{name}: {target_str}[/yellow]")
-                else:
-                    console.print(f"  [red]{name}: {target_str}[/red]")
-            else:
-                # Weekly target block
-                target_str = f"{completed}/{weekly}"
-                if completed >= weekly:
-                    console.print(f"  [green]{name}: {target_str}[/green]")
-                elif completed >= weekly // 2:
-                    console.print(f"  [yellow]{name}: {target_str}[/yellow]")
-                else:
-                    console.print(f"  [red]{name}: {target_str}[/red]")
-
-    # === TASKS SECTION ===
-    # Get completed tasks this week
-    completed_tasks = get_tasks_completed_since(week_start_str)
-
-    # Get pending tasks
-    pending = get_tasks(status="pending")
-    overdue = get_overdue_tasks()
-    overdue_ids = {t["id"] for t in overdue}
-
-    # Summary counts
-    console.print(f"\n[bold]Tasks[/bold]")
-    console.print(f"  Completed this week: [green]{len(completed_tasks)}[/green]")
-    console.print(f"  Still pending:       [yellow]{len(pending)}[/yellow]")
-    if overdue:
-        console.print(f"  Overdue:             [red]{len(overdue)}[/red]")
-
-    # Completed tasks list
-    if completed_tasks:
-        console.print(f"\n[bold green]Completed This Week ({len(completed_tasks)})[/bold green]")
-        for t in completed_tasks:
-            completed_dt = datetime.fromisoformat(t["completed_at"])
-            day_str = completed_dt.strftime("%a")
-            ws_str = f" [dim]({t['workstream']})[/dim]" if t["workstream"] else ""
-            console.print(f"  [green][/green] {t['title']}{ws_str} [dim]{day_str}[/dim]")
-    else:
-        console.print(f"\n[dim]No tasks completed this week yet[/dim]")
-
-    # Pending tasks list
-    if pending:
-        console.print(f"\n[bold yellow]Still To Do ({len(pending)})[/bold yellow]")
-        for t in pending:
-            due_str = ""
-            if t["due_date"]:
-                if t["id"] in overdue_ids:
-                    due_str = f" [red](overdue: {t['due_date']})[/red]"
-                elif t["due_date"] == today.isoformat():
-                    due_str = f" [yellow](due today)[/yellow]"
-                else:
-                    due_str = f" [dim](due {t['due_date']})[/dim]"
-
-            pri_marker = ""
-            if t["priority"] == "high":
-                pri_marker = "[red]![/red] "
-
-            ws_str = f" [dim]({t['workstream']})[/dim]" if t["workstream"] else ""
-            console.print(f"  {pri_marker}{t['title']}{ws_str}{due_str}")
-    else:
-        console.print(f"\n[green]All tasks complete![/green]")
-
-    console.print()
-
-
-@cli.command("demo-setup")
-def demo_setup():
-    """Set up demo data for showcasing the tool. Only works in demo mode (UE_DEMO=1)."""
-    from datetime import datetime, timedelta
-    from ue.config import DEMO_MODE, DATA_DIR, CONFIG_PATH
-    from ue.db import get_db, add_task, set_block_target, log_block_completion
-
-    if not DEMO_MODE:
-        console.print("[red]Demo setup only works in demo mode.[/red]")
-        console.print("Run with: [cyan]UE_DEMO=1 ue demo-setup[/cyan]")
-        return
-
-    console.print(f"[bold]Setting up demo data in {DATA_DIR}[/bold]\n")
-
-    # Create demo config
-    demo_config = {
-        "workstreams": {
-            "work": {"priority": "high", "color": "green"},
-            "side-project": {"priority": "mid", "color": "blue"},
-            "learning": {"priority": "mid", "color": "yellow"},
-            "health": {"priority": "low", "color": "cyan"}
-        },
-        "git_repos": [],
-        "watch_dirs": [],
-        "last_sync": datetime.now().isoformat()
-    }
-    import json
-    DATA_DIR.mkdir(parents=True, exist_ok=True)
-    CONFIG_PATH.write_text(json.dumps(demo_config, indent=2))
-    console.print("  Created config with workstreams")
-
-    # Initialize DB
-    db = get_db()
-    db.close()
-
-    # Create block targets
-    blocks = [
-        ("Deep Work", 5, "work"),
-        ("Exercise", 4, "health"),
-        ("Reading", 3, "learning"),
-        ("Side Project", 2, "side-project"),
-        ("Meditation", 0, "health"),  # daily
-    ]
-    for name, target, ws in blocks:
-        set_block_target(name, target, ws)
-    console.print(f"  Created {len(blocks)} block targets")
-
-    # Add some block completions for this week
-    today = datetime.now().date()
-    week_start = today - timedelta(days=today.weekday())
-
-    completions = [
-        ("Deep Work", week_start, "completed"),
-        ("Deep Work", week_start + timedelta(days=1), "completed"),
-        ("Exercise", week_start, "completed"),
-        ("Exercise", week_start + timedelta(days=1), "completed"),
-        ("Reading", week_start, "completed"),
-        ("Side Project", week_start, "completed"),
-    ]
-    for name, date, status in completions:
-        log_block_completion(name, date.isoformat(), status)
-    console.print(f"  Logged {len(completions)} block completions")
-
-    # Create tasks
-    tasks = [
-        ("Review PR from teammate", (today + timedelta(days=1)).isoformat(), "work", "high"),
-        ("Write blog post draft", (today + timedelta(days=3)).isoformat(), "side-project", "normal"),
-        ("Update project README", (today + timedelta(days=5)).isoformat(), "side-project", "low"),
-        ("Schedule dentist appointment", (today + timedelta(days=2)).isoformat(), None, "normal"),
-        ("Research new framework", None, "learning", "low"),
-    ]
-    for title, due, ws, pri in tasks:
-        add_task(title, due, ws, pri)
-    console.print(f"  Created {len(tasks)} tasks")
-
-    # Mark one task as done
-    from ue.db import complete_task
-    complete_task(1)  # Mark first task done
-    console.print("  Marked 1 task as completed")
-
-    console.print("\n[green]Demo setup complete![/green]")
-    console.print("\nRun demo commands with:")
-    console.print("  [cyan]UE_DEMO=1 ue status[/cyan]")
-    console.print("  [cyan]UE_DEMO=1 ue am[/cyan]")
-    console.print("  [cyan]UE_DEMO=1 ue did[/cyan]")
-    console.print("  [cyan]UE_DEMO=1 ue done[/cyan]")
-    console.print("  [cyan]UE_DEMO=1 ue task add[/cyan]")
-
-
-@cli.command("demo-reset")
-def demo_reset():
-    """Reset demo data. Only works in demo mode (UE_DEMO=1)."""
-    import shutil
-    from ue.config import DEMO_MODE, DATA_DIR
-
-    if not DEMO_MODE:
-        console.print("[red]Demo reset only works in demo mode.[/red]")
-        console.print("Run with: [cyan]UE_DEMO=1 ue demo-reset[/cyan]")
-        return
-
-    if DATA_DIR.exists():
-        shutil.rmtree(DATA_DIR)
-        console.print(f"[yellow]Removed {DATA_DIR}[/yellow]")
-    else:
-        console.print("[dim]Demo directory doesn't exist[/dim]")
-
-    console.print("\nRun [cyan]UE_DEMO=1 ue demo-setup[/cyan] to recreate demo data")
+cli.add_command(task)
+cli.add_command(done)
+
+# Block commands
+cli.add_command(block)
+cli.add_command(did)
+
+# Routine commands
+cli.add_command(am)
+cli.add_command(pm)
+cli.add_command(review)
+cli.add_command(status)
+cli.add_command(focus)
+
+# Log/mark commands
+cli.add_command(log)
+cli.add_command(mark)
+
+# Demo commands
+cli.add_command(demo_setup)
+cli.add_command(demo_reset)
 
 
 if __name__ == "__main__":
