+++ ue/commands/__init__.py
@@ -1,0 +1,1 @@
+"""CLI command modules."""
